+++ src/instruments/option.rs
@@ -91,6 +91,10 @@ impl FinancialInstrument for EuropeanOption {
         Some(self.expiry)
     }
 
+    fn clearing(&self) -> Option<Clearing> {
+        Some(self.clearing)
+    }
+
     fn instrument_type(&self) -> &str {
         "EuropeanOption"
     }